backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
websockets>=12.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-huggingface>=0.1.0
langchain-chroma>=0.1.0
langgraph>=0.2.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
pdfplumber>=0.11.0
keybert>=0.8.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
